+AppController blank class
This thing is basically an instance version of the main entry point--we'll scoop out everything from Program.Main and put it in here. This is huge because we can manage the lifecycle of this class, especially for testing and dependency delivery. This is way harder with a static class like Program.